ナンバリング登録APIを別ファイルに切り出した
