+++ index.js
@@ -659,85 +659,7 @@ sequelize.sync().then(() => {
     // あるユーザのナンバリングを取得
     // あるステッカーのナンバリングについて、全ユーザの分布を調べたりするためには、
     // 別のAPIが必要
-    app.get(process.env.EXPRESS_ROOT + '/numbering/:part/:userName', (req, res, next) => {
-        const part = req.params.part;
-        const userName = req.params.userName;
-        const letters = req.query.letters;
-
-        if (!userName || !(part === 'corner' || part === 'edgeMiddle')) {
-            logger.emit('api.request', {
-                requestType: 'GET',
-                endpoint: '/hinemos/numbering/',
-                params: {
-                    part,
-                    userName,
-                },
-                status: 'error',
-                code: 400,
-                msg: '',
-            });
-
-            res.status(400).send(badRequestError);
-            return;
-        }
-
-        const query = {
-            where: {
-                userName,
-            },
-        };
-        if (letters) {
-            query.where.letter = letters.split(/(.)/).filter(x => x);
-        }
-
-        // cornerかedgeMiddle以外の場合、既にハジかれている
-        let numberingModel;
-        if (part === 'corner') {
-            numberingModel = NumberingCorner;
-        } else if (part === 'edgeMiddle') {
-            numberingModel = NumberingEdgeMiddle;
-        }
-
-        return numberingModel
-            .findAll(query)
-            .then((result) => {
-                logger.emit('api.request', {
-                    requestType: 'GET',
-                    endpoint: '/hinemos/numbering/',
-                    params: {
-                        part,
-                        userName,
-                    },
-                    status: 'success',
-                    code: 200,
-                    msg: '',
-                });
-
-                const ans = {
-                    success: {
-                        code: 200,
-                        result,
-                    },
-                };
-                res.json(ans);
-                res.status(200);
-            })
-            .catch(() => {
-                logger.emit('api.request', {
-                    requestType: 'GET',
-                    endpoint: '/hinemos/numbering/',
-                    params: {
-                        part,
-                        userName,
-                    },
-                    status: 'error',
-                    code: 400,
-                    msg: '',
-                });
-
-                res.status(400).send(badRequestError);
-            });
-    });
+    app.get(process.env.EXPRESS_ROOT + '/numbering/:part/:userName', route.numbering.getProcess);
 
     app.get(process.env.EXPRESS_ROOT + '/threeStyleQuizLog/:part/:userName', (req, res, next) => {
         const userName = req.params.userName;
@@ -1228,116 +1150,7 @@ sequelize.sync().then(() => {
     app.post(`${process.env.EXPRESS_ROOT}/threeStyle/:part`, route.threeStyle.postProcess);
 
     // 部分更新することは考えていないので、そのユーザのデータを全置き換え
-    app.post(process.env.EXPRESS_ROOT + '/numbering/:part/', (req, res, next) => {
-        const part = req.params.part;
-        const userName = req.decoded.userName;
-        const numberings = Array.from(new Set(req.body.numberings)); // [{sticker, numbering,}]
-
-        // ナンバリングで重複を排除した時に数が一致しないということは、重複が存在するということなのでNG
-        const uniqedLn = Array.from(new Set(numberings.map(x => x.letter))).length;
-        const assertCond = uniqedLn === numberings.length;
-
-        if (!userName || !req.body.numberings || uniqedLn === 0 || !assertCond || !(part === 'corner' || part === 'edgeMiddle')) {
-            logger.emit('api.request', {
-                requestType: 'POST',
-                endpoint: '/hinemos/numbering/',
-                params: {
-                    part,
-                    userName,
-                    numberings,
-                    decoded: req.decoded,
-                },
-                status: 'error',
-                code: 400,
-                msg: '',
-            });
-
-            res.status(400).send(badRequestError);
-            return;
-        }
-
-        let numberingModel;
-        if (part === 'corner') {
-            numberingModel = NumberingCorner;
-        } else if (part === 'edgeMiddle') {
-            numberingModel = NumberingEdgeMiddle;
-        }
-
-        sequelize
-            .transaction((t) => {
-                // まず今のnumberingを消す
-                return numberingModel
-                    .destroy({
-                        where: {
-                            userName,
-                        },
-                        transaction: t,
-                    })
-                    .then((result) => {
-                        // 次に、UIから入力された情報で更新
-                        const promises = [];
-                        for (let i = 0; i < numberings.length; i++) {
-                            const sticker = numberings[i].sticker;
-                            const letter = numberings[i].letter;
-                            const instance = {
-                                userName,
-                                sticker,
-                                letter,
-                            };
-
-                            promises.push(
-                                numberingModel
-                                    .create(instance, {
-                                        transaction: t,
-                                    })
-                                    .then((result) => {
-                                        return {
-                                            code: 200,
-                                            params: instance,
-                                            msg: 'OK',
-                                        };
-                                    }));
-                        }
-
-                        return Promise.all(promises)
-                            .then((ans) => {
-                                logger.emit('api.request', {
-                                    requestType: 'POST',
-                                    endpoint: '/hinemos/numbering/',
-                                    params: {
-                                        part,
-                                        userName,
-                                        numberings,
-                                        decoded: req.decoded,
-                                    },
-                                    status: 'success',
-                                    code: 200,
-                                    msg: '',
-                                });
-
-                                res.json(ans);
-                                res.status(200);
-                            })
-                            .catch((err) => {
-                                logger.emit('api.request', {
-                                    requestType: 'POST',
-                                    endpoint: '/hinemos/numbering/',
-                                    params: {
-                                        part,
-                                        userName,
-                                        numberings,
-                                        decoded: req.decoded,
-                                    },
-                                    status: 'erorr',
-                                    code: 400,
-                                    msg: err,
-                                });
-
-                                res.status(400).send(badRequestError);
-                            });
-                    });
-            });
-    });
+    app.post(process.env.EXPRESS_ROOT + '/numbering/:part/', route.numbering.postProcess);
 
     app.post(process.env.EXPRESS_ROOT + '/threeStyleQuizLog/:part', (req, res, next) => {
         const userName = req.decoded.userName;
